requirements.txt
# LangChain Core
langchain
langchain-core

# Document loaders
langchain_community
PyPDF

# Parsing tool
bs4

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# To print the chain graph
grandalf

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn